6-size.c

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of char: %lu bytes\n", (unsigned long)sizeof(a));
-	printf("Size of int: %lu bytes\n", (unsigned long)sizeof(b));
-	printf("Size of long: %lu bytes\n", (unsigned long)sizeof(c));
-	printf("Size of long long int: %lu bytes\n", (unsigned long)sizeof(d));
-	printf("Size of float: %lu bytes\n", (unsigned long)sizeof(e));
+	printf("Size of char: %lu bytes\n", sizeof(a));
+	printf("Size of int: %lu bytes\n", sizeof(b));
+	printf("Size of long: %lu bytes\n", sizeof(c));
+	printf("Size of long long int: %lu bytes\n", sizeof(d));
+	printf("Size of float: %lu bytes\n", sizeof(e));
 	return (0);
 }
